Added PostPolicy

+++ app/Http/Controllers/PostController.php
@@ -18,7 +18,8 @@ class PostController extends Controller
         // $posts = Post::get();
 
         // user and likes relationship
-        $posts = Post::with(['user', 'likes'])->paginate(20);
+        // $posts = Post::orderBy('created_at', 'desc')->with(['user', 'likes'])->paginate(20);
+        $posts = Post::latest()->with(['user', 'likes'])->paginate(20);
 
         return view('posts.index', [
             'posts' => $posts
@@ -48,7 +49,10 @@ class PostController extends Controller
     }
 
     public function destroy(Post $post, Request $request) {
-        
+
+        // 'delete' is the method name inside "PostPolicy"
+        $this->authorize('delete', $post);
+
         $post->delete();
 
         return back();
